Forked the Crowd Process fork of formula.js
This is a straight-forward fork of the Crowd Process fork of Sutoiku's
formula.js. The Crowd Process fork is notable primarily in that it
features significantly improved error handling, cuts down on
dependencies and excludes functions which are not offered by Microsoft
Excel itself. This fork uses Webpack (like the original formula.js)
and features a current GitHub release, enabling it to be used with the
dependency management system Bower.