Fixed the unit strings accepted by the CONVERT function
The unit strings accepted by the CONVERT engineering function have
been improved. Some units, such as atomic units and natural units, did
not have complete unit symbols before, partly because some of them use
the same unit symbol (such as "a.u. of action" and "n.u. of action,"
which both use "ħ"). These units used question marks in place of
usable unit symbols. This commit instead introduces usable symbols.
Some of them are not in use (and have essentially been made up for the
sake of this fork of formulajs), and documentation will thus be
required in order for users to be able to make use of them.