Fixed a large number of bugs related to date, time and math functions
A large number of bugs have been fixed related to the date and time
functions as well as one function related to the math and trigonometry
category. While this work impacts many functions, NETWORKDAYS,
NETWORKDAYS.INTL, WORKDAY and CEILING in particular have received
attention.

The test suite that is part of this project has unfortunately not been
updated to take the new changes into account. There is a comprehensive
test suite testing the new changes, but at this point in time, that
test suite is maintained outside of this project. It would be
desirable to improve the formulajs test suite in the future.

In detail:
* All dates are forced to be UTC dates (thereby negating the Date
  class's tendency to make use of the local time zone when creating
  dates). This solves a large number of bugs with the date and time
  functions.
* When converting Excel sequential serial numbers to JavaScript time
  stamps, the number is rounded prematurely. This mitigates
  floating-point issues that would cause date calculations to fail
  (returning the wrong day, for instance).
* utils.parseDate() no longer assumes that Excel sequential serial
  numbers do not have a fractional component (representing time). This
  fixes a large number of bugs related to time functions.
* The behavior of NETWORKDAYS and NETWORKDAYS.INTL now matches that of
  Excel. This entails returning a number one greater than the one
  previously returned.
* WORKDAY has seen various improvements:
  * Like Excel, its days parameter can now be negative, enabling the
    clock to be set back as well as advanced.
  * It now always an integer, that is, without a fractional component
    denoting time.
  * Issues with daylight saving time have been fixed.
* CEILING now correctly handles significance parameters which are
  negative.