Fixed rounding issues with CEILING, FLOOR and related functions
The CEILING, CEILING.MATH, FLOOR and FLOOR.MATH functions sometimes
prematurely rounded numbers. The implementation used the ROUND
function and erroneously converted the significance to a number of
digits, as required by the ROUND function. This new implementation
sidesteps the issue by using the MROUND function instead, which needs
no such conversion process.

This project is tested by a comprehensive test suite, which
unfortunately is maintained externally. For the record, the new
implementation returns 0.675 when CEILING(0.65625, 0.025) is invoked,
whereas the old implementation returned 0.68.